+++ src/concurrent/logbuffer/term_scan.rs
@@ -191,7 +191,7 @@ mod tests {
         let aligned_message_length = bit_utils::align(message_length, frame_descriptor::FRAME_ALIGNMENT);
 
         // Write message length in to log
-        log.put_ordered::<i32>(frame_descriptor::length_offset(offset), message_length as i32);
+        log.put_ordered::<i32>(frame_descriptor::length_offset(offset), message_length);
 
         // Set type
         log.put::<u16>(frame_descriptor::type_offset(offset), data_frame_header::HDR_TYPE_DATA);
@@ -214,11 +214,11 @@ mod tests {
         let aligned_message_length = bit_utils::align(message_length, frame_descriptor::FRAME_ALIGNMENT);
 
         // Write message length in to log
-        log.put_ordered::<i32>(frame_descriptor::length_offset(offset), message_length as i32);
+        log.put_ordered::<i32>(frame_descriptor::length_offset(offset), message_length);
         // Set type
         log.put::<u16>(frame_descriptor::type_offset(offset), data_frame_header::HDR_TYPE_DATA);
         // Write next message length
-        log.put_ordered::<i32>(frame_descriptor::length_offset(aligned_message_length), message_length as i32);
+        log.put_ordered::<i32>(frame_descriptor::length_offset(aligned_message_length), message_length);
         // Set type
         log.put::<u16>(
             frame_descriptor::type_offset(aligned_message_length),
@@ -244,11 +244,11 @@ mod tests {
         let third_message_length = limit_offset - (2 * aligned_message_length);
 
         // Write message length in to log
-        log.put_ordered::<i32>(frame_descriptor::length_offset(offset), message_length as i32);
+        log.put_ordered::<i32>(frame_descriptor::length_offset(offset), message_length);
         // Set type
         log.put::<u16>(frame_descriptor::type_offset(offset), data_frame_header::HDR_TYPE_DATA);
         // Write next message length
-        log.put_ordered::<i32>(frame_descriptor::length_offset(aligned_message_length), message_length as i32);
+        log.put_ordered::<i32>(frame_descriptor::length_offset(aligned_message_length), message_length);
         // Set type for second msg
         log.put::<u16>(
             frame_descriptor::type_offset(aligned_message_length),
@@ -257,7 +257,7 @@ mod tests {
         // Write next message length
         log.put_ordered::<i32>(
             frame_descriptor::length_offset(aligned_message_length * 2),
-            third_message_length as i32,
+            third_message_length,
         );
         // Set type for third msg
         log.put::<u16>(
@@ -281,11 +281,11 @@ mod tests {
         let limit_offset = (2 * aligned_message_length) + 1;
 
         // Write message length in to log
-        log.put_ordered::<i32>(frame_descriptor::length_offset(offset), message_length as i32);
+        log.put_ordered::<i32>(frame_descriptor::length_offset(offset), message_length);
         // Set type
         log.put::<u16>(frame_descriptor::type_offset(offset), data_frame_header::HDR_TYPE_DATA);
         // Write next message length as 0 in to log
-        log.put_ordered::<i32>(frame_descriptor::length_offset(aligned_message_length), message_length as i32);
+        log.put_ordered::<i32>(frame_descriptor::length_offset(aligned_message_length), message_length);
         // Set type for second msg
         log.put::<u16>(
             frame_descriptor::type_offset(aligned_message_length),
@@ -294,7 +294,7 @@ mod tests {
         // Write next message length as 0 in to log
         log.put_ordered::<i32>(
             frame_descriptor::length_offset(aligned_message_length * 2),
-            message_length as i32,
+            message_length,
         );
         // Set type for second msg
         log.put::<u16>(
@@ -318,7 +318,7 @@ mod tests {
         let limit_offset = aligned_message_length - 1;
 
         // Write message length in to log
-        log.put_ordered::<i32>(frame_descriptor::length_offset(offset), message_length as i32);
+        log.put_ordered::<i32>(frame_descriptor::length_offset(offset), message_length);
         // Set type
         log.put::<u16>(frame_descriptor::type_offset(offset), data_frame_header::HDR_TYPE_DATA);
 
@@ -338,7 +338,7 @@ mod tests {
         let limit_offset = aligned_message_length;
 
         // Write message length in to log
-        log.put_ordered::<i32>(frame_descriptor::length_offset(offset), message_length as i32);
+        log.put_ordered::<i32>(frame_descriptor::length_offset(offset), message_length);
         // Set type
         log.put::<u16>(frame_descriptor::type_offset(offset), data_frame_header::HDR_TYPE_DATA);
 
@@ -358,11 +358,11 @@ mod tests {
         let limit_offset = log.capacity();
 
         // Write message length in to log
-        log.put_ordered::<i32>(frame_descriptor::length_offset(offset), message_length as i32);
+        log.put_ordered::<i32>(frame_descriptor::length_offset(offset), message_length);
         // Set type
         log.put::<u16>(frame_descriptor::type_offset(offset), data_frame_header::HDR_TYPE_DATA);
         // Write message length in to log
-        log.put_ordered::<i32>(frame_descriptor::length_offset(aligned_message_length), message_length as i32);
+        log.put_ordered::<i32>(frame_descriptor::length_offset(aligned_message_length), message_length);
         // Set type
         log.put::<u16>(
             frame_descriptor::type_offset(aligned_message_length),
@@ -388,7 +388,7 @@ mod tests {
 
         log.set_memory(0, log.capacity(), 0);
 
-        log.put_ordered::<i32>(frame_offset, data_frame_header::LENGTH as i32);
+        log.put_ordered::<i32>(frame_offset, data_frame_header::LENGTH);
 
         let handler = |term_id: i32, _buffer: &AtomicBuffer, offset: Index, length: Index| {
             assert_eq!(TERM_ID, term_id);
@@ -415,12 +415,12 @@ mod tests {
 
         log.put_ordered::<i32>(
             tail - frame_descriptor::ALIGNED_HEADER_LENGTH,
-            data_frame_header::LENGTH as i32,
+            data_frame_header::LENGTH,
         );
         log.put_ordered::<i32>(tail, 0);
         log.put_ordered::<i32>(
             high_water_mark - frame_descriptor::ALIGNED_HEADER_LENGTH,
-            data_frame_header::LENGTH as i32,
+            data_frame_header::LENGTH,
         );
 
         let handler = |term_id: i32, _buffer: &AtomicBuffer, offset: Index, length: Index| {
@@ -448,12 +448,12 @@ mod tests {
 
         log.put_ordered::<i32>(
             tail - frame_descriptor::ALIGNED_HEADER_LENGTH,
-            data_frame_header::LENGTH as i32,
+            data_frame_header::LENGTH,
         );
         log.put_ordered::<i32>(tail, 0);
         log.put_ordered::<i32>(
             high_water_mark - frame_descriptor::ALIGNED_HEADER_LENGTH,
-            data_frame_header::LENGTH as i32,
+            data_frame_header::LENGTH,
         );
 
         let handler = |term_id: i32, _buffer: &AtomicBuffer, offset: Index, length: Index| {
@@ -480,7 +480,7 @@ mod tests {
 
         log.set_memory(0, log.capacity(), 0);
 
-        log.put_ordered::<i32>(tail, padding_length as i32);
+        log.put_ordered::<i32>(tail, padding_length);
         log.put_ordered::<i32>(tail + data_frame_header::LENGTH, 0);
 
         let handler = |_term_id: i32, _buffer: &AtomicBuffer, _offset: Index, _length: Index| {
@@ -502,8 +502,8 @@ mod tests {
 
         let scan_outcome = term_scan::scan_outcome(padding, available);
 
-        assert_eq!(padding as i32, term_scan::padding(scan_outcome));
-        assert_eq!(available as i32, term_scan::available(scan_outcome));
+        assert_eq!(padding, term_scan::padding(scan_outcome));
+        assert_eq!(available, term_scan::available(scan_outcome));
     }
 
     #[test]
@@ -516,13 +516,13 @@ mod tests {
         let aligned_frame_length = bit_utils::align(frame_length, frame_descriptor::FRAME_ALIGNMENT);
         let frame_offset = 0;
 
-        log.put_ordered::<i32>(frame_offset, frame_length as i32);
+        log.put_ordered::<i32>(frame_offset, frame_length);
         log.put::<u16>(frame_descriptor::type_offset(frame_offset), data_frame_header::HDR_TYPE_DATA);
         log.put_ordered::<i32>(aligned_frame_length, 0);
 
         let scan_outcome = term_scan::scan_for_availability(&log, frame_offset, MTU_LENGTH);
 
-        assert_eq!(aligned_frame_length as i32, term_scan::available(scan_outcome));
+        assert_eq!(aligned_frame_length, term_scan::available(scan_outcome));
         assert_eq!(0, term_scan::padding(scan_outcome));
     }
 
@@ -537,7 +537,7 @@ mod tests {
         let max_length = aligned_frame_length - 1;
         let frame_offset = 0;
 
-        log.put_ordered::<i32>(frame_offset, frame_length as i32);
+        log.put_ordered::<i32>(frame_offset, frame_length);
         log.put::<u16>(frame_descriptor::type_offset(frame_offset), data_frame_header::HDR_TYPE_DATA);
 
         let scan_outcome = term_scan::scan_for_availability(&log, frame_offset, max_length);
@@ -557,10 +557,10 @@ mod tests {
         let aligned_length_one = bit_utils::align(frame_length_one, frame_descriptor::FRAME_ALIGNMENT);
         let aligned_length_two = bit_utils::align(frame_length_two, frame_descriptor::FRAME_ALIGNMENT);
 
-        buffer.put_ordered::<i32>(frame_offset, aligned_length_one as i32);
+        buffer.put_ordered::<i32>(frame_offset, aligned_length_one);
         buffer.put::<u16>(frame_descriptor::type_offset(frame_offset), frame_type_one);
 
-        buffer.put_ordered::<i32>(frame_offset + aligned_length_one, frame_length_two as i32);
+        buffer.put_ordered::<i32>(frame_offset + aligned_length_one, frame_length_two);
         buffer.put::<u16>(
             frame_descriptor::type_offset(frame_offset + aligned_length_one),
             frame_type_two,
@@ -589,7 +589,7 @@ mod tests {
 
         let scan_outcome = term_scan::scan_for_availability(&log, 0, MTU_LENGTH);
 
-        assert_eq!(total_length as i32, term_scan::available(scan_outcome));
+        assert_eq!(total_length, term_scan::available(scan_outcome));
         assert_eq!(0, term_scan::padding(scan_outcome));
     }
 
@@ -613,7 +613,7 @@ mod tests {
 
         let scan_outcome = term_scan::scan_for_availability(&log, frame_offset, MTU_LENGTH);
 
-        assert_eq!(total_length as i32, term_scan::available(scan_outcome));
+        assert_eq!(total_length, term_scan::available(scan_outcome));
         assert_eq!(0, term_scan::padding(scan_outcome));
     }
 
@@ -637,7 +637,7 @@ mod tests {
 
         let scan_outcome = term_scan::scan_for_availability(&log, frame_offset, MTU_LENGTH);
 
-        assert_eq!(frame_one_length as i32, term_scan::available(scan_outcome));
+        assert_eq!(frame_one_length, term_scan::available(scan_outcome));
         assert_eq!(0, term_scan::padding(scan_outcome));
     }
 
@@ -649,12 +649,12 @@ mod tests {
         let aligned_frame_length = bit_utils::align(data_frame_header::LENGTH * 2, frame_descriptor::FRAME_ALIGNMENT);
         let frame_offset = LOG_BUFFER_CAPACITY - aligned_frame_length;
 
-        log.put_ordered::<i32>(frame_offset, aligned_frame_length as i32);
+        log.put_ordered::<i32>(frame_offset, aligned_frame_length);
         log.put::<u16>(frame_descriptor::type_offset(frame_offset), data_frame_header::HDR_TYPE_DATA);
 
         let scan_outcome = term_scan::scan_for_availability(&log, frame_offset, MTU_LENGTH);
 
-        assert_eq!(aligned_frame_length as i32, term_scan::available(scan_outcome));
+        assert_eq!(aligned_frame_length, term_scan::available(scan_outcome));
         assert_eq!(0, term_scan::padding(scan_outcome));
     }
 
@@ -709,7 +709,7 @@ mod tests {
 
         let scan_outcome = term_scan::scan_for_availability(&log, frame_offset, mtu);
 
-        assert_eq!(aligned_frame_length as i32, term_scan::available(scan_outcome));
+        assert_eq!(aligned_frame_length, term_scan::available(scan_outcome));
         assert_eq!(0, term_scan::padding(scan_outcome));
     }
 }
